requirements.txt
fastapi
uvicorn
transformers
torch
pandas
python-dotenv
scikit-learn
requests
streamlit
